Variable Explore: Refactor QLineEdit for filtering and custom proxy
