+++ spyder/preferences/shortcuts.py
@@ -13,9 +13,8 @@ import re
 # Third party imports
 from qtpy import PYQT5
 from qtpy.compat import from_qvariant, to_qvariant
-from qtpy.QtCore import (QAbstractTableModel, QModelIndex, QRegExp,
-                         QSortFilterProxyModel, Qt, Slot, QEvent)
-from qtpy.QtGui import (QKeySequence, QRegExpValidator, QIcon)
+from qtpy.QtCore import QAbstractTableModel, QModelIndex, Qt, Slot, QEvent
+from qtpy.QtGui import QKeySequence, QIcon
 from qtpy.QtWidgets import (QAbstractItemView, QApplication, QDialog,
                             QGridLayout, QHBoxLayout, QLabel,
                             QLineEdit, QMessageBox, QPushButton, QSpacerItem,
@@ -29,9 +28,10 @@ from spyder.config.gui import (get_shortcut, iter_shortcuts,
 from spyder.preferences.configdialog import GeneralConfigPage
 from spyder.utils import icon_manager as ima
 from spyder.utils.qthelpers import get_std_icon, create_toolbutton
-from spyder.utils.stringmatching import get_search_scores, get_search_regex
-from spyder.widgets.helperwidgets import HTMLDelegate
-from spyder.widgets.helperwidgets import HelperToolButton
+from spyder.utils.stringmatching import get_search_scores
+from spyder.widgets.helperwidgets import (CustomSortFilterProxy,
+                                          FinderLineEdit, HelperToolButton,
+                                          HTMLDelegate, VALID_FINDER_CHARS)
 
 
 # Valid shortcut keys
@@ -40,11 +40,6 @@ EDITOR_SINGLE_KEYS = SINGLE_KEYS + ["Home", "End", "Ins", "Enter",
                                     "Return", "Backspace", "Tab",
                                     "PageUp", "PageDown", "Clear",  "Pause",
                                     "Left", "Up", "Right", "Down"]
-
-# Valid finder chars. To be improved
-VALID_ACCENT_CHARS = "ÁÉÍOÚáéíúóàèìòùÀÈÌÒÙâêîôûÂÊÎÔÛäëïöüÄËÏÖÜñÑ"
-VALID_FINDER_CHARS = r"[A-Za-z\s{0}]".format(VALID_ACCENT_CHARS)
-
 
 # Key sequences blacklist for the shortcut editor dialog
 BLACKLIST = {}
@@ -118,29 +113,11 @@ class ShortcutLineEdit(QLineEdit):
         super(ShortcutLineEdit, self).setText(sequence)
 
 
-class ShortcutFinder(QLineEdit):
+class ShortcutFinder(FinderLineEdit):
     """Textbox for filtering listed shortcuts in the table."""
 
-    def __init__(self, parent, callback=None):
-        super(ShortcutFinder, self).__init__(parent)
-        self._parent = parent
-
-        # Widget setup
-        regex = QRegExp(VALID_FINDER_CHARS + "{100}")
-        self.setValidator(QRegExpValidator(regex))
-
-        # Signals
-        if callback:
-            self.textChanged.connect(callback)
-
-    def set_text(self, text):
-        """Set the filter text."""
-        text = text.strip()
-        new_text = self.text() + text
-        self.setText(new_text)
-
     def keyPressEvent(self, event):
-        """Qt Override."""
+        """Qt and FilterLineEdit Override."""
         key = event.key()
         if key in [Qt.Key_Up]:
             self._parent.previous_row()
@@ -646,38 +623,6 @@ class ShortcutsModel(QAbstractTableModel):
         """"Reset model to take into account new search letters."""
         self.beginResetModel()
         self.endResetModel()
-
-
-class CustomSortFilterProxy(QSortFilterProxyModel):
-    """Custom column filter based on regex."""
-
-    def __init__(self, parent=None):
-        super(CustomSortFilterProxy, self).__init__(parent)
-        self._parent = parent
-        self.pattern = re.compile(r'')
-
-    def set_filter(self, text):
-        """Set regular expression for filter."""
-        self.pattern = get_search_regex(text)
-        if self.pattern:
-            self._parent.setSortingEnabled(False)
-        else:
-            self._parent.setSortingEnabled(True)
-        self.invalidateFilter()
-
-    def filterAcceptsRow(self, row_num, parent):
-        """Qt override.
-
-        Reimplemented from base class to allow the use of custom filtering.
-        """
-        model = self.sourceModel()
-        name = model.row(row_num).name
-        r = re.search(self.pattern, name)
-
-        if r is None:
-            return False
-        else:
-            return True
 
 
 class ShortcutsTable(QTableView):
